requirements.txt
pymongo
cryptography
python-dotenv